Миграцией колонка status в таблице заказов была переименована в status_id, сделан индекс и удалённый ключ к таблице order_statuses

+++ migrations/m210721_095420_modify_users_table_unique.php
@@ -14,7 +14,7 @@ class m210721_095420_modify_users_table_unique extends Migration
     {
         $this->alterColumn('users', 'login', $this->string()->unique());
         $this->alterColumn('users', 'email', $this->string()->unique());
-        $this->alterColumn('users', 'phone', $this->integer()->unique());
+        $this->alterColumn('users', 'phone', $this->bigInteger()->unique());
     }
 
     /**
